clerk webhook complete

+++ app/api/webhook/clerk/route.js
@@ -11,20 +11,50 @@ if (!CLERK_WEBHOOK_SECRET) {
   console.error("❌ CLERK_WEBHOOK_SECRET is undefined");
 }
 
+function getName(firstName, lastName) {
+  if (firstName && lastName) {
+    return `${firstName} ${lastName}`;
+  } else if (firstName) {
+    return firstName;
+  } else if (lastName) {
+    return lastName;
+  }
+
+  return "Unknown";
+}
+
 export async function POST(req) {
   try {
-    const { body } = await req.json();
+    const svix_id = req.headers.get("svix-id") ?? "";
+    const svix_timestamp = req.headers.get("svix-timestamp") ?? "";
+    const svix_signature = req.headers.get("svix-signature") ?? "";
 
-    console.log("🔹 Webhook received:", body);
+    const resBody = await req.text();
+
+    const sivx = new Webhook(CLERK_WEBHOOK_SECRET);
+
+    const body = sivx.verify(resBody, {
+      "svix-id": svix_id,
+      "svix-timestamp": svix_timestamp,
+      "svix-signature": svix_signature,
+    });
 
     const { type, data } = body;
+    const user = await db
+      .select()
+      .from(users)
+      .where(eq(users.clerkUserId, data.id));
 
     if (type === "user.created") {
+      if (user.length > 0) {
+        return NextResponse.json(
+          { message: "User already exists" },
+          { status: 400 }
+        );
+      }
+
       await db.insert(users).values({
-        name:
-          data.first_name && data.last_name
-            ? `${data.first_name} ${data.last_name}`
-            : "Unknown",
+        name: getName(data.first_name, data.last_name),
         email: data.email_addresses[0].email_address,
         clerkUserId: data.id,
         createdAt: new Date(data.created_at),
@@ -33,13 +63,17 @@ export async function POST(req) {
 
       console.log("✅ User created:", data.email_addresses[0].email_address);
     } else if (type === "user.updated") {
+      if (user.length == 0) {
+        return NextResponse.json(
+          { message: "User does not exist" },
+          { status: 404 }
+        );
+      }
+
       await db
         .update(users)
         .set({
-          name:
-            data.first_name && data.last_name
-              ? `${data.first_name} ${data.last_name}`
-              : "Unknown",
+          name: getName(data.first_name, data.last_name),
           email: data.email_addresses[0].email_address,
           updatedAt: new Date(data.updated_at),
         })
@@ -47,6 +81,13 @@ export async function POST(req) {
 
       console.log("🔄 User updated:", data.email_addresses[0].email_address);
     } else if (type === "user.deleted") {
+      if (user.length == 0) {
+        return NextResponse.json(
+          { message: "User does not exist" },
+          { status: 404 }
+        );
+      }
+
       await db.delete(users).where(eq(users.clerkUserId, data.id));
 
       console.log("❌ User deleted:", data.id);
